Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main()
 {
-    int a,b,c;
-    scanf("%d%d%d",&a,&b,&c);
+    float a,b,c;
+    scanf("%f%f%f",&a,&b,&c);
     average=(a+b+c)/3;
-    printf("Average:%f\n",average,a,b,c);
+    printf("Average:%.2f\n",average,a,b,c);
     return 0;
 }
